Test servisi eklendi

+++ UniTrack.Application/Abstraction/Repositories/ICommentRepository.cs
@@ -10,8 +10,8 @@ namespace UniTrack.Application.Abstraction.Repositories
         public Task<List<Comment>> GetAllCommentsByClubIdAsync(Guid clubId);
         public Task<List<Comment>> GetAllCommentByEventIdAsync(Guid eventId);
         public Task<List<Comment>> GetAllCommentsByUserIdAsync(Guid userId);
-        public Task<float> GetClubAverageRatingAsync(Guid clubId);
-        public Task<float> GetEventAverageRatingAsync(Guid eventId);
+        public Task<double> GetClubAverageRatingAsync(Guid clubId);
+        public Task<double> GetEventAverageRatingAsync(Guid eventId);
         Task<Dictionary<Guid, (float AverageRating, int Count)>> GetEventsRatingsSummaryAsync(List<Guid> eventIds);
     }
 }
